feat: moved cards.json to src, tested website, started to refactor the TS
